Move types to common file
