+++ plugins/woocommerce-beta-tester/src/product-editor-dev-tools/product-editor-dev-tools-bar.tsx
@@ -154,7 +154,7 @@ export function ProductEditorDevToolsBar( {
 						isSelected={ selectedTab === 'product' }
 					/>
 					<TemplateTabPanel
-						postType={ postType }
+						evaluationContext={ evaluationContext }
 						isSelected={ selectedTab === 'template' }
 					/>
 					<TemplateEventsTabPanel
