+++ plugins/woocommerce-beta-tester/src/product-editor-dev-tools/template-tab-panel/block-template-details-panel.tsx
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { BlockTemplate } from './types';
+import { BlockTemplate } from '../types';
 
 export function BlockTemplateDetailsPanel( {
 	blockTemplate,
