+++ plugins/woocommerce-beta-tester/src/product-editor-dev-tools/types.ts
@@ -1,3 +1,13 @@
+/**
+ * External dependencies
+ */
+import { Product } from '@woocommerce/data';
+
+export interface EvaluationContext {
+	postType: string;
+	editedProduct: Product;
+}
+
 export interface BlockTemplateAttributes {
 	[ key: string ]: unknown;
 	_templateBlockId: string;
